sdg benchmark flag to skip data write to disk

+++ source/standalone_examples/benchmarks/benchmark_sdg.py
@@ -51,6 +51,7 @@ parser.add_argument(
     choices=["LocalLogMetrics", "JSONFileMetrics", "OsmoKPIFile"],
     help="Benchmarking backend, defaults",
 )
+parser.add_argument("--skip-write", action="store_true", help="Skip writing annotator data to disk")
 
 
 args, unknown = parser.parse_known_args()
@@ -64,6 +65,7 @@ delete_data_when_done = args.delete_data_when_done
 print_results = args.print_results
 headless = args.headless
 n_gpu = args.num_gpus
+skip_write = args.skip_write
 
 if "all" in args.annotators:
     annotators_kwargs = {annotator: True for annotator in VALID_ANNOTATORS}
@@ -80,6 +82,7 @@ print(f"\tdisable_viewport_rendering: {disable_viewport_rendering}")
 print(f"\tdelete_data_when_done: {delete_data_when_done}")
 print(f"\print_results: {print_results}")
 print(f"\theadless: {headless}")
+print(f"\tskip_write: {skip_write}")
 
 import os
 import shutil
@@ -140,14 +143,22 @@ for i in range(num_cameras):
 render_products = []
 for i, cam in enumerate(cameras):
     render_products.append(rep.create.render_product(cam, (width, height), name=f"rp_{i}"))
-writer = rep.writers.get("BasicWriter")
-output_directory = (
-    os.getcwd()
-    + f"/_out_sdg_benchmark_{num_frames}_frames_{num_cameras}_cameras_{asset_count}_asset_count_{len(annotators_kwargs)}_annotators"
-)
-print(f"[SDG Benchmark] Output directory: {output_directory}")
-writer.initialize(output_dir=output_directory, **annotators_kwargs)
-writer.attach(render_products)
+if skip_write:
+    print("[SDG Benchmark] Skipping writing to disk, attaching annotators to render products..")
+    for annot_type, enabled in annotators_kwargs.items():
+        if enabled:
+            annot = rep.AnnotatorRegistry.get_annotator(annot_type)
+            for rp in render_products:
+                annot.attach(rp)
+else:
+    writer = rep.writers.get("BasicWriter")
+    output_directory = (
+        os.getcwd()
+        + f"/_out_sdg_benchmark_{num_frames}_frames_{num_cameras}_cameras_{asset_count}_asset_count_{len(annotators_kwargs)}_annotators"
+    )
+    print(f"[SDG Benchmark] Output directory: {output_directory}")
+    writer.initialize(output_dir=output_directory, **annotators_kwargs)
+    writer.attach(render_products)
 assets = rep.create.group([cubes, cones, cylinders, spheres, tori])
 cameras = rep.create.group(cameras)
 
@@ -171,14 +182,17 @@ for _ in range(10):
     omni.kit.app.get_app().update()
 benchmark.store_measurements()
 
-benchmark.set_phase("benchmark")
 print("[SDG Benchmark] Starting SDG..")
+benchmark.set_phase("benchmark")
 start_time = time.time()
 rep.orchestrator.run_until_complete(num_frames=num_frames)
 end_time = time.time()
+benchmark.store_measurements()
+omni.kit.app.get_app().update()
+
 duration = end_time - start_time
 avg_frametime = duration / num_frames
-if delete_data_when_done:
+if delete_data_when_done and not skip_write:
     print(f"[SDG Benchmark] Deleting data: {output_directory}")
     shutil.rmtree(output_directory)
 if print_results:
@@ -187,8 +201,6 @@ if print_results:
     print(f"[SDG Benchmark] avg FPS: {1 / avg_frametime:.2f}")
     results_csv = f"{num_frames}, {num_cameras}, {width}, {height}, {asset_count}, {duration:.4f}, {avg_frametime:.4f}, {1 / avg_frametime:.2f}"
     print(f"num_frames, num_cameras, width, height, asset_count, duration, avg_frametime, avg_fps\n{results_csv}\n")
-omni.kit.app.get_app().update()
-benchmark.store_measurements()
 
 benchmark.stop()
